Fixes #1
The bug was actually trivial, after knowing what to look for ;P
Interleave will stop merging the collections at the smallest
collection, and since we’re partitioning the deck it would never return
all of the collection.

I’m sure there’s a better way to do it, but I just added a helper
method to pad out the collections. Suggestions are welcome :)